pyproject.toml
[project]
name = "ssis-to-pyspark"
version = "2.0.0"
description = "LLM-powered SSIS to PySpark migration accelerator"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@company.com" }
]
keywords = ["ssis", "sql-server", "pyspark", "migration", "etl", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # LLM Providers
    "ollama>=0.4.0",
    "google-generativeai>=0.8.0",
    "anthropic>=0.40.0",
    # Core
    "pyyaml>=6.0",
    "lxml>=5.0",
    "rich>=13.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    # Logging & Monitoring
    "structlog>=24.0.0",
    # Config Management
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyspark>=4.1.0",
    "jupyter>=1.1.1",
    "ipykernel>=7.1.0",
    "openai>=2.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[project.scripts]
ssis2spark = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
filterwarnings = [
    "ignore::FutureWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
